requirements.txt
rich
google-cloud-compute
google-cloud-monitoring
pandas